move router engine info into a debug resource behind an env variable
