+++ graylog2-server/src/main/java/org/graylog2/rest/resources/RestResourcesModule.java
@@ -65,6 +65,7 @@ import org.graylog2.rest.resources.system.UrlWhitelistResource;
 import org.graylog2.rest.resources.system.contentpacks.CatalogResource;
 import org.graylog2.rest.resources.system.contentpacks.ContentPackResource;
 import org.graylog2.rest.resources.system.debug.DebugEventsResource;
+import org.graylog2.rest.resources.system.debug.DebugStreamsResource;
 import org.graylog2.rest.resources.system.indexer.FailuresResource;
 import org.graylog2.rest.resources.system.indexer.IndexSetDefaultsResource;
 import org.graylog2.rest.resources.system.indexer.IndexSetsResource;
@@ -106,6 +107,7 @@ public class RestResourcesModule extends Graylog2Module {
         addLegacySearchResources();
         addProcessingResources();
         addStreamsResources();
+        addDebugResources();
 
         addSystemRestResource(HelloWorldResource.class);
         addSystemRestResource(RolesResource.class);
@@ -126,6 +128,13 @@ public class RestResourcesModule extends Graylog2Module {
         addSystemRestResource(TrafficResource.class);
         addSystemRestResource(SearchVersionResource.class);
         addSystemRestResource(EntityListPreferencesResource.class);
+    }
+
+    private void addDebugResources() {
+        if(Boolean.parseBoolean(System.getenv("GRAYLOG_ENABLE_DEBUG_RESOURCES"))) {
+            // TODO: move the DebugEventsResource under this env property check as well?
+            addSystemRestResource(DebugStreamsResource.class);
+        }
     }
 
     private void addAuthResources() {
